style: renaming style files
