+++ src/shared/components/ToolsTech/styles.module.scss
